+++ packages/gamut-labs/src/AppHeader/AppHeaderElements/AppHeaderLinkSections/index.tsx
@@ -1,3 +1,4 @@
+import { ContentContainer } from '@codecademy/gamut';
 import { css, states } from '@codecademy/gamut-styles';
 import styled from '@emotion/styled';
 import React from 'react';
@@ -20,6 +21,7 @@ export type AppHeaderLinkSectionsProps = {
   onKeyDown?: (event: React.KeyboardEvent) => void;
   onFocus?: () => void;
   onBlur?: () => void;
+  mobile?: boolean;
 };
 
 type LinkComponentProps = {
@@ -78,28 +80,35 @@ const LinkComponent: React.FC<LinkComponentProps> = ({
 export const AppHeaderLinkSections = React.forwardRef<
   HTMLUListElement,
   AppHeaderLinkSectionsProps
->(({ action, item, showIcon = false, onKeyDown, ...props }, ref) => (
-  <StyledList ref={ref} {...props}>
-    {item.type === 'profile-dropdown'
-      ? item.popover.map((linkSection: AppHeaderLinkItem[], sectionIndex) =>
-          linkSection.map((link: AppHeaderLinkItem, linkIndex) => (
-            <LinkComponent
-              onKeyDown={onKeyDown}
-              key={link.id}
-              action={action}
-              link={link}
-              showLineBreak={sectionIndex !== 0 && linkIndex === 0}
-              showIcon={showIcon}
-            />
-          ))
-        )
-      : item.popover.map((link: AppHeaderLinkItem) => (
-          <LinkComponent
-            onKeyDown={onKeyDown}
-            key={link.id}
-            action={action}
-            link={link}
-          />
-        ))}
-  </StyledList>
-));
+>(
+  (
+    { action, item, showIcon = false, onKeyDown, mobile = false, ...props },
+    ref
+  ) => (
+    <ContentContainer size={mobile ? 'medium' : 'small'}>
+      <StyledList ref={ref} {...props}>
+        {item.type === 'profile-dropdown'
+          ? item.popover.map((linkSection: AppHeaderLinkItem[], sectionIndex) =>
+              linkSection.map((link: AppHeaderLinkItem, linkIndex) => (
+                <LinkComponent
+                  onKeyDown={onKeyDown}
+                  key={link.id}
+                  action={action}
+                  link={link}
+                  showLineBreak={sectionIndex !== 0 && linkIndex === 0}
+                  showIcon={showIcon}
+                />
+              ))
+            )
+          : item.popover.map((link: AppHeaderLinkItem) => (
+              <LinkComponent
+                onKeyDown={onKeyDown}
+                key={link.id}
+                action={action}
+                link={link}
+              />
+            ))}
+      </StyledList>
+    </ContentContainer>
+  )
+);