+++ packages/gamut-labs/src/GlobalFooter/FooterNavLinks/CompanyLinks.tsx
@@ -87,6 +87,10 @@ export const CompanyLinks: React.FC<CompanyLinksProps> = ({
             Learner Stories
           </Anchor>
         </FooterLinkItem>
+        {/* Refer a friend marketing anchor */}
+        <FooterLinkItem>
+          <span id="extole_zone_global_footer" />
+        </FooterLinkItem>
       </FooterLinkItems>
     </Box>
   );